refactor: extract ScreenshotImageWriter.swift from ScreenshotCaptureSupport.swift
Splits ScreenshotCaptureSupport.swift into two files by concern:
- ScreenshotCaptureSupport.swift keeps display-capture types
  (ScreenCaptureDisplaySnapshot, CapturedDisplayImage,
  ScreenCaptureImageProviding, ScreenCaptureKitImageProvider,
  and its `+AppError` extension)
- ScreenshotImageWriter.swift gets the PNG-writing types
  (ScreenshotImageWriting protocol + PNGScreenshotImageWriter)

Byte-identical move of lines 138-184. New file imports CoreGraphics,
Foundation, ImageIO, UniformTypeIdentifiers (needed for
CGImageDestination + UTType). ScreenshotCaptureSupport.swift keeps its
CoreGraphics + Foundation + @preconcurrency ScreenCaptureKit.

ScreenshotCaptureSupport.swift: 184 → 136 lines. Tests: 667.
